Crate holochain_wasm_utils

Library holding necessary code for the Ribosome that is also useful for hdk-rust, or more generally for making rust code that the Ribosome can run. Must not have any dependency with any other Holochain crates.

Re-exports

pub extern crate holochain_core_types;

pub extern crate holochain_json_derive;

pub extern crate holochain_json_api;

pub extern crate holochain_persistence_api;

Modules

api_serialization
ignore api_serialization because it is nothing but structs to hold serialization

macros

memory

Macros

core_error
Macro for creating a CoreError from a HolochainError on the spot with file!() and line!()

core_error_generic
Macro for creating a generic CoreError on the spot with file!() and line!()

ribosome_error_code
Macro for creating a RibosomeErrorCode as a RuntimeValue Result-Option on the spot

ribosome_success

zome_assert
Macro for creating a RibosomeErrorCode as a RuntimeValue Result-Option on the spot Will panic! if out or memory or other serialization error occured.

Functions

wasm_target_dir